fixing properties multiple assign problem
